# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Documentation Title",
        "description": "This is the API documentation for the news aggregator service.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Fetch a paginated list of articles with filters",
                "description": "Retrieve articles with optional filters for keyword, date, category, and source.",
                "operationId": "11aa4cc6561a7b94c56d1201d62513ee",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Search for articles with a keyword in the title or content",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "‘Naked and Afraid’ star Sarah Danser dies in crash at age 34 - New York Post"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Filter articles by a specific publication date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-10-25 08:07:00"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter articles by category",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": ""
                        }
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "description": "Filter articles by source",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "NewsAPI Org"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A paginated list of articles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Latest in Tech"
                                                    },
                                                    "content": {
                                                        "type": "string",
                                                        "example": "An overview of technology trends..."
                                                    },
                                                    "author": {
                                                        "type": "string",
                                                        "example": "Jane Doe"
                                                    },
                                                    "source": {
                                                        "type": "string",
                                                        "example": "BBC News"
                                                    },
                                                    "category": {
                                                        "type": "string",
                                                        "example": "Technology"
                                                    },
                                                    "published_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-10-25T08:07:00Z"
                                                    },
                                                    "image_url": {
                                                        "type": "string",
                                                        "example": "https://example.com/image.jpg"
                                                    },
                                                    "url": {
                                                        "type": "string",
                                                        "example": "https://example.com/article"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "http://localhost:8000/api/articles?page=1"
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "last_page_url": {
                                            "type": "string",
                                            "example": "http://localhost:8000/api/articles?page=5"
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 50
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid filter values provided"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/articles/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Fetch a specific article by ID",
                "description": "Retrieve details of a particular article using its unique ID.",
                "operationId": "getArticle",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the article to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Article Title"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "This is the content of the article."
                                        },
                                        "author": {
                                            "type": "string",
                                            "example": "Author Name"
                                        },
                                        "source": {
                                            "type": "string",
                                            "example": "News Source"
                                        },
                                        "category": {
                                            "type": "string",
                                            "example": "Technology"
                                        },
                                        "published_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-10-25T08:07:00"
                                        },
                                        "image_url": {
                                            "type": "string",
                                            "example": "https://example.com/image.jpg"
                                        },
                                        "url": {
                                            "type": "string",
                                            "example": "https://example.com/article"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Article not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
...
}